refactor(bus): split the Bus trait impl into accessors.rs
Move the impl crate::Bus for SystemBus block (~430 lines — the hot-path
read_u8/16/32 + write_u8/16/32 with routing + bit-band translation) out of
bus/mod.rs into bus/accessors.rs. Trait impls may live in any file in the
defining crate, so external callers are unaffected.

bus/mod.rs 2467 -> 2038 (3509 -> 2038 across the 3 bus PRs: from_config, tick,
accessors). Verified both feature sets: workspace lib + jit,event-scheduler lib
+ firmware_survival (26) + register_coverage, clippy clean.
